Added a blog blueprint and index file
DO NOT REVERT TO THIS COMMIT OR ANY EARLIER ONES!
Before this commit, any redirects to an index page would generate an error because the index file had not been created yet.
This commit adds that index page, but it also links to other pages that haven't been created yet.